check if connected in the controllers / Put Owner done / Dto done

+++ Controllers/CaveController.cs
@@ -3,6 +3,7 @@ using CaveManager.Entities.DTO;
 using CaveManager.Repository;
 using CaveManager.Repository.Repository.Contract;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace CaveManager.Controllers
 {
@@ -20,49 +21,76 @@ namespace CaveManager.Controllers
             _logger = logger;
         }
 
+        [NonAction]
+        public bool IsConnected()
+        {
+            var identity = User?.Identity as ClaimsIdentity;
+            var idCurrentUser = identity?.FindFirst(ClaimTypes.NameIdentifier);
+            if (idCurrentUser == null)
+                return true;
+            else
+                return false;
+        }
+
         /// <summary>
-        /// select Cave with his id
+        /// Select Cave with his id
         /// </summary>
         /// <param name="idCave"></param>
         /// <returns></returns>
         [HttpGet("{idCave}")]
         public async Task<ActionResult<Cave>> GetCave(int idCave)
         {
-            var cave = await caveRepository.SelectCaveAsync(idCave);
-            if (cave != null)
-                return Ok(cave);
-            else
-                return BadRequest("Cave not found");
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var cave = await caveRepository.SelectCaveAsync(idCave);
+                if (cave != null)
+                    return Ok(cave);
+                else
+                    return BadRequest("Cave not found");
+            }
+            return BadRequest("Not logged");
         }
+
         /// <summary>
-        /// Select All Cave of a Owner with his id
+        /// Select All Owner's cave(s) with his id
         /// </summary>
         /// <param name="idOwner"></param>
         /// <returns></returns>
         [HttpGet("{idOwner}")]
         public async Task<ActionResult<List<Cave>>> GetAllCaveFromOwner(int idOwner)
         {
-            var caves = await caveRepository.GetAllCaveFromAOwner(idOwner);
-            if (caves != null)
-                return Ok(caves);
-            else
-                return BadRequest("Owner not found");
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var caves = await caveRepository.GetAllCaveFromAOwner(idOwner);
+                if (caves != null)
+                    return Ok(caves);
+                else
+                    return BadRequest("Owner not found");
+            }
+            return BadRequest("Not logged");
         }
+
         /// <summary>
-        /// Select all Drawer of a Cave with his id
+        /// Select all Cave's drawner(s) with his id
         /// </summary>
         /// <param name="idCave"></param>
         /// <returns></returns>
         [HttpGet("{idCave}")]
         public async Task<ActionResult<List<Drawer>>> GetAllDrawer(int idCave)
         {
-            var drawers = await caveRepository.GetAllDrawerFromACave(idCave);
-            if (drawers!= null)
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                return Ok(drawers);
+                var drawers = await caveRepository.GetAllDrawerFromACave(idCave);
+                if (drawers != null)
+                {
+                    return Ok(drawers);
+                }
+                return BadRequest("No Cave found");
             }
-            return BadRequest("No Cave found");
-            
+            return BadRequest("Not logged");
         }
 
         /// <summary>
@@ -74,12 +102,17 @@ namespace CaveManager.Controllers
         [HttpPost("{idOwner}")]
         public async Task<ActionResult<Cave>> AddCave(DTOCave dtoCave, int idOwner)
         {
-            var cave = new Cave { Name = dtoCave.Name, OwnerId = idOwner};
-            var caveCreated = await caveRepository.AddCaveAsync(cave);
-            if (caveCreated != null)
-                return Ok(caveCreated);
-            else
-                return BadRequest("Owner not found");
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var cave = new Cave { Name = dtoCave.Name, OwnerId = idOwner };
+                var caveCreated = await caveRepository.AddCaveAsync(cave);
+                if (caveCreated != null)
+                    return Ok(caveCreated);
+                else
+                    return BadRequest("Owner not found");
+            }
+            return BadRequest("Not logged");
         }
 
         /// <summary>
@@ -92,15 +125,18 @@ namespace CaveManager.Controllers
         [HttpPut("{id}")]
         public async Task<ActionResult<Cave>> UpdateCave(int id, string name)
         {
-            var portedevoiture = await caveRepository.UpdateCaveAsync(id, name);
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var portedevoiture = await caveRepository.UpdateCaveAsync(id, name);
 
-            if (portedevoiture != null)
-                return Ok(portedevoiture);
-            else
-                return BadRequest("Cave not found");
-
+                if (portedevoiture != null)
+                    return Ok(portedevoiture);
+                else
+                    return BadRequest("Cave not found");
+            }
+            return BadRequest("Not logged");
         }
-
 
         /// <summary>
         /// Delete a Cave
@@ -110,18 +146,18 @@ namespace CaveManager.Controllers
         [HttpDelete("{idCave}")]
         public async Task<ActionResult<Cave>> RemoveCave(int idCave)
         {
-            var deleteCave = await caveRepository.RemoveCaveAsync(idCave);
-            if (deleteCave != null)
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                return Ok(deleteCave);
-
+                var deleteCave = await caveRepository.RemoveCaveAsync(idCave);
+                if (deleteCave != null)
+                {
+                    return Ok(deleteCave);
+                }
+                return BadRequest("Cave not found");
             }
-            return BadRequest("Cave not found");
-
+            return BadRequest("Not logged");
         }
-
-
-
     }
 }
