+++ Controllers/DrawerController.cs
@@ -3,7 +3,7 @@ using CaveManager.Entities.DTO;
 using CaveManager.Repository;
 using CaveManager.Repository.Repository.Contract;
 using Microsoft.AspNetCore.Mvc;
-
+using System.Security.Claims;
 
 namespace CaveManager.Controllers
 {
@@ -16,11 +16,22 @@ namespace CaveManager.Controllers
         private readonly ILogger<DrawerController> _logger;
         public DrawerController(IDrawerRepository drawerRepository, ILogger<DrawerController> logger, IWebHostEnvironment environment)
         {
-
             this.drawerRepository = drawerRepository;
             this.environment = environment;
             _logger = logger;
         }
+
+        [NonAction]
+        public bool IsConnected()
+        {
+            var identity = User?.Identity as ClaimsIdentity;
+            var idCurrentUser = identity?.FindFirst(ClaimTypes.NameIdentifier);
+            if (idCurrentUser == null)
+                return true;
+            else
+                return false;
+        }
+
         /// <summary>
         /// Select a drawer with his id
         /// </summary>
@@ -29,10 +40,15 @@ namespace CaveManager.Controllers
         [HttpGet("{idDrawer}")]
         public async Task<ActionResult<Drawer>> GetDrawer(int idDrawer)
         {
-            var drawers = await drawerRepository.SelectDrawerAsync(idDrawer);
-
-            return Ok(drawers);
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var drawers = await drawerRepository.SelectDrawerAsync(idDrawer);
+                return Ok(drawers);
+            }
+            return BadRequest("Not logged");
         }
+
         /// <summary>
         /// Get wines from a drawer
         /// </summary>
@@ -41,25 +57,41 @@ namespace CaveManager.Controllers
         [HttpGet("{idDrawer}")]
         public async Task<ActionResult<List<Wine>>> GetAllWinesFromADrawer(int idDrawer)
         {
-            var wines = await drawerRepository.GetAllWinesFromADrawerAsync(idDrawer);
-            if (wines != null)
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                return Ok(wines);
+                var wines = await drawerRepository.GetAllWinesFromADrawerAsync(idDrawer);
+                if (wines != null)
+                {
+                    return Ok(wines);
+                }
+                return BadRequest("Drawers not found");
             }
-            return BadRequest("Drawers not found");
-
+            return BadRequest("Not logged");
         }
+
+        /// <summary>
+        /// Add a drawer
+        /// </summary>
+        /// <param name="dTODrawer"></param>
+        /// <returns></returns>
         [HttpGet("{idDrawer}")]
         public async Task<ActionResult<DTODrawer>> AddDrawer(DTODrawer dTODrawer)
         {
-            var nouveauDrawer = new Drawer { Name = dTODrawer.Name, MaxPlace = dTODrawer.MaxPlace, PlaceUsed = dTODrawer.PlaceUsed };
-            var drawer = await drawerRepository.AddDrawerAsync(nouveauDrawer);
-            if (drawer != null)
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                return Ok(dTODrawer);
+                var nouveauDrawer = new Drawer { Name = dTODrawer.Name, MaxPlace = dTODrawer.MaxPlace, PlaceUsed = dTODrawer.PlaceUsed };
+                var drawer = await drawerRepository.AddDrawerAsync(nouveauDrawer);
+                if (drawer != null)
+                {
+                    return Ok(dTODrawer);
+                }
+                return BadRequest();
             }
-            return BadRequest();
+            return BadRequest("Not logged");
         }
+
         /// <summary>
         /// Update a Drawer with his id, change all the data
         /// </summary>
@@ -69,14 +101,20 @@ namespace CaveManager.Controllers
         [HttpPut("{idDrawer}")]
         public async Task<ActionResult<Drawer>> UpdateDrawer(int idDrawer, DTODrawer dTODrawer)
         {
-            var nouveauDrawer = new Drawer {Name=dTODrawer.Name, MaxPlace=dTODrawer.MaxPlace, PlaceUsed=dTODrawer.PlaceUsed };
-            var drawer = await drawerRepository.UpdateDrawerAsync(idDrawer, nouveauDrawer);
-            if (drawer != null)
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                return Ok(drawer);
+                var nouveauDrawer = new Drawer { Name = dTODrawer.Name, MaxPlace = dTODrawer.MaxPlace, PlaceUsed = dTODrawer.PlaceUsed };
+                var drawer = await drawerRepository.UpdateDrawerAsync(idDrawer, nouveauDrawer);
+                if (drawer != null)
+                {
+                    return Ok(drawer);
+                }
+                return BadRequest("Drawer not found");
             }
-            return BadRequest("Drawer not found");
+            return BadRequest("Not logged");
         }
+
         /// <summary>
         /// Delete a Drawer with his id
         /// </summary>
@@ -85,12 +123,17 @@ namespace CaveManager.Controllers
         [HttpDelete("{idCave}")]
         public async Task<ActionResult<Drawer>> DeleteDrawer(int drawerId)
         {
-            var drawer = await drawerRepository.RemoveDrawerAsync(drawerId);
-            if (drawer != null)
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                return Ok(drawer);
+                var drawer = await drawerRepository.RemoveDrawerAsync(drawerId);
+                if (drawer != null)
+                {
+                    return Ok(drawer);
+                }
+                return BadRequest("Drawer not found");
             }
-            return BadRequest("Drawer not found");
+            return BadRequest("Not logged");
         }
 
     }
