+++ Controllers/WineController.cs
@@ -23,6 +23,17 @@ namespace CaveManager.Controllers
             _logger = logger;
         }
 
+        [NonAction]
+        public bool IsConnected()
+        {
+            var identity = User?.Identity as ClaimsIdentity;
+            var idCurrentUser = identity?.FindFirst(ClaimTypes.NameIdentifier);
+            if (idCurrentUser == null)
+                return true;
+            else
+                return false;
+        }
+
         /// <summary>
         /// Add a Wine to Drawer
         /// </summary>
@@ -32,23 +43,26 @@ namespace CaveManager.Controllers
         [HttpPost("{idDrawer}")]
         public async Task<ActionResult<Wine>> AddWine(int idDrawer, [FromForm] DTOWine dTOwine)
         {
-            var wine = new Wine
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                Name = dTOwine.Name,
-                Type = dTOwine.Type,
-                Designation = dTOwine.Designation,
-                Bottling = dTOwine.Bottling,
-                MaxVintageRecommended = dTOwine.MaxVintageRecommended,
-                MinVintageRecommended = dTOwine.MinVintageRecommended
-            };
-            var wineAdded = await wineRepository.AddWineAsync(wine, idDrawer);
+                var wine = new Wine
+                {
+                    Name = dTOwine.Name,
+                    Type = dTOwine.Type,
+                    Designation = dTOwine.Designation,
+                    Bottling = dTOwine.Bottling,
+                    MaxVintageRecommended = dTOwine.MaxVintageRecommended,
+                    MinVintageRecommended = dTOwine.MinVintageRecommended
+                };
+                var wineAdded = await wineRepository.AddWineAsync(wine, idDrawer);
 
-            if (wineAdded.error == "ok")
-                return Ok(wineAdded.wine);
-            else
-                return BadRequest(wineAdded.error);
-
-            return BadRequest("not logged");
+                if (wineAdded.error == "ok")
+                    return Ok(wineAdded.wine);
+                else
+                    return BadRequest(wineAdded.error);
+            }
+            return BadRequest("Not logged");
         }
 
 
@@ -60,9 +74,14 @@ namespace CaveManager.Controllers
         [HttpGet("{idWine}")]
         public async Task<ActionResult<Wine>> GetWine(int idWine)
         {
-            var wine = await wineRepository.GetWineAsync(idWine);
-            if (wine != null) return Ok(wine);
-            return BadRequest("Wine not found");
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var wine = await wineRepository.GetWineAsync(idWine);
+                if (wine != null) return Ok(wine);
+                return BadRequest("Wine not found");
+            }
+            return BadRequest("Not logged");
         }
 
         /// <summary>
@@ -73,9 +92,14 @@ namespace CaveManager.Controllers
         [HttpDelete("{idWine}")]
         public async Task<ActionResult<Wine>> DeleteWine(int idWine)
         {
-            var wine = await wineRepository.DeleteWineAsync(idWine);
-            if (wine != null) return Ok(wine);
-            return BadRequest("Wine not found, try with a correct id");
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var wine = await wineRepository.DeleteWineAsync(idWine);
+                if (wine != null) return Ok(wine);
+                return BadRequest("Wine not found, try with a correct id");
+            }
+            return BadRequest("Not logged");
         }
 
         /// <summary>
@@ -91,14 +115,18 @@ namespace CaveManager.Controllers
         [HttpPut("{idWine}")]
         public async Task<ActionResult<Wine>> PutWine(int idWine, DTOWine dTOWine)
         {
-            var newDrawer = new Wine { Name = dTOWine.Name, Type = dTOWine.Type, Designation = dTOWine.Designation, Bottling = dTOWine.Bottling, MaxVintageRecommended = dTOWine.MaxVintageRecommended, MinVintageRecommended = dTOWine.MinVintageRecommended };
-            var wine = await wineRepository.PutWineAsync(newDrawer, idWine);
-            if (wine != null)
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
             {
-                return Ok(wine);
+                var newDrawer = new Wine { Name = dTOWine.Name, Type = dTOWine.Type, Designation = dTOWine.Designation, Bottling = dTOWine.Bottling, MaxVintageRecommended = dTOWine.MaxVintageRecommended, MinVintageRecommended = dTOWine.MinVintageRecommended };
+                var wine = await wineRepository.PutWineAsync(newDrawer, idWine);
+                if (wine != null)
+                {
+                    return Ok(wine);
+                }
+                return BadRequest("Wine not found, try with a correct id");
             }
-            return BadRequest("Wine not found, try with a correct id");
-
+            return BadRequest("Not logged");
         }
 
         /// <summary>
@@ -110,11 +138,16 @@ namespace CaveManager.Controllers
         [HttpPost("{idDrawer}")]
         public async Task<ActionResult<Wine>> DuplicateWine(int idWine, int idDrawer)
         {
-            var wine = await wineRepository.DuplicateWineAsync(idWine, idDrawer);
-            if (wine.error != "ok")
-                return Ok(wine.wine);
-            else
-                return BadRequest(wine.error);
+            bool checkIsConnected = IsConnected();
+            if (checkIsConnected)
+            {
+                var wine = await wineRepository.DuplicateWineAsync(idWine, idDrawer);
+                if (wine.error != "ok")
+                    return Ok(wine.wine);
+                else
+                    return BadRequest(wine.error);
+            }
+            return BadRequest("Not logged");
         }
     }
 }
